requirements.txt
pandas
plotly
python-dotenv
dash
dash_bootstrap_components
gunicorn
requests